hjdivad/index -> hjdivad/init
Neovim will automatically check for foo/init when requiring foo. See
:help lua-require.
