+++ home/.config/nvim.symlink/lua/hjdivad/init.lua
@@ -1,3 +1,105 @@
+---Show the nvim-tree
+---
+---If the tree is not open, open it, otherwise switch to the tree's window.
+---
+---If the current buffer has a name, also find it in the tree.
+---
+---*note* finding files works poorly for those outside of `$CWD`
+local function toggle_nvim_tree()
+  -- TODO: make this work better when opening a file outside of <cwd>, as well as opening a file within <cwd> after nvimtree has cd-d outside
+  --  i.e. NVIMTreeOpen dirname (file) or cwd + findfile
+  -- TODO: make this work for NEW buffers (i.e buffers never saved)
+  local buffer_name = vim.fn['bufname']()
+
+  if buffer_name == '' then
+    vim.cmd('NvimTreeOpen')
+  else
+    vim.cmd('NvimTreeFindFile')
+  end
+end
+
+---Open a REPL terminal that runs `repl_cmd` and begin editing the REPL input buffer `".repl." .. repl_extension`.
+---The terminal is placed top-left if any neoterm instances are open, and vertically leftmost otherwise.
+---`:write`ing to the RPEL file will send the contents to the repl. The extension is used for file type detection.
+---
+---@see setup_vimtest()
+---@param repl_cmd string a command to start a REPL in a neoterm instance
+---@param repl_extension string the name of the file to use for input commands to the REPL
+local function edit_repl(repl_cmd, repl_extension)
+  local get_neoterm_window_ids = require('hjdivad/terminal').get_neoterm_window_ids
+  -- TODO: make a toggle_repl that closes the repl windows if they're open
+  local repl_file = '.repl.' .. repl_extension
+  if not vim.g.neoterm.repl or not vim.g.neoterm.repl.instance_id then
+    local starting_window = vim.api.nvim_get_current_win()
+    -- create the REPL buffer
+    local open_terminals = get_neoterm_window_ids()
+    if #open_terminals > 0 then
+      -- TODO: open the terminal if it's not visible
+      -- if we have a terminal open on the left, create the REPL buffer in the topleft
+      vim.cmd([[
+        " Create a new neoterm window in the top-left
+        100wincmd h
+        100wincmd k
+        above new
+        Tnew
+        normal G
+      ]])
+    else
+      -- if we have no open terminals, create one vertically on the left
+      vim.cmd([[
+        vertical topleft Tnew
+        normal G
+      ]])
+    end
+
+    -- start the REPL
+    vim.cmd([[exe g:neoterm.last_id . 'T ]] .. repl_cmd .. "'")
+    -- Set the new neoterm as the REPL terminal
+    vim.cmd([[
+      exe 'TREPLSetTerm ' . g:neoterm.last_id
+    ]])
+
+    -- return to our starting window to edit the repl input file
+    vim.api.nvim_set_current_win(starting_window)
+  end
+
+  vim.cmd('edit ' .. repl_file)
+  vim.cmd('stopinsert')
+end
+
+local function edit_generic_repl()
+  local repl_cmd = nil
+  local repl_extension = nil
+
+  vim.ui.input({ prompt = 'REPL command (e.g. node) > ', default = 'node' },
+    function(input) repl_cmd = input end)
+
+  vim.ui.input({ prompt = 'REPL input file extension (e.g. js) > ', default = 'js' },
+    function(input) repl_extension = input end)
+
+  if not repl_cmd then error('No REPL command given. Please specify a commnad to start the REPL.') end
+
+  if not repl_extension then
+    error(
+      'No extension given. Please specify an extension (to trigger filetype) for your REPL input buffer')
+  end
+
+  edit_repl(repl_cmd, repl_extension)
+
+  local suggested_keymap = [[vim.api.nvim_set_keymap('n', '<leader>re', function() require('hjdivad').edit_repl(']] .. repl_cmd .. [[', ']] .. repl_extension .. [[') end]]
+  vim.notify([[To skip inputs next time, consider adding `]] .. suggested_keymap ..
+    [[` to .vimrc.lua]], vim.log.levels.INFO, {})
+end
+
+local function setup_repls()
+  vim.cmd([[
+    augroup REPL
+      autocmd!
+      autocmd BufWritePre .repl.* exe 'TREPLSendFile'
+    augroup end
+  ]])
+end
+
 local function setup_local_config()
   ---spelling
   vim.opt.spelllang = { 'sv', 'en_gb', 'en_us' }
@@ -5,6 +107,9 @@ local function setup_local_config()
     '.vimspell.utf8.add', '~/.local/share/.nvim/spell/en.utf-8.add',
     '~/.local/share/.nvim/spell/sv.utf-8.add'
   }
+
+  -- inform terminals that they are within an nvim instance
+  vim.env.NVIM_WRAPPER = 1
 end
 
 local function setup_clipboard()
@@ -74,11 +179,9 @@ local function setup_key_mappings()
 
   vim.keymap.set('n', '<leader><leader>', '<cmd>nohl | checktime<cr>', { desc = 'use ,, to clear highlights', })
 
-  vim.keymap.set('n', '<leader>nf', function()
-    require('hjdivad/index').toggle_nvim_tree()
-  end, { desc = 'now files (toggle nvim-tree)', })
+  vim.keymap.set('n', '<leader>nf', toggle_nvim_tree, { desc = 'now files (toggle nvim-tree)', })
   vim.keymap.set('n', '<leader>nt', function()
-    require('hjdivad/index').toggle_terminal()
+    require('hjdivad/terminal').toggle_terminal()
   end, { desc = 'now terminal (intelligent neoterm toggling)', })
 
   vim.keymap.set('n', '<leader>ff', function() require('telescope.builtin').find_files() end, { desc = 'find files relative to `cwd`', })
@@ -124,30 +227,30 @@ local function setup_key_mappings()
   vim.keymap.set('n', '<leader>ts', '<Cmd>Telescope tmux windows<cr>', { desc = 'tmux: select window', })
   vim.keymap.set('n', '<leader>tt', '<Cmd>silent !tmux switch-client -l<cr>', { desc = 'tmux: toggle previous window', })
   vim.keymap.set('n', '<leader>td', function()
-    require('hjdivad/index').goto_tmux_session('todos', 'todos')
+    require('hjdivad/tmux').goto_tmux_session('todos', 'todos')
   end, { desc = 'tmux: todos', })
   vim.keymap.set('n', '<leader>tr', function()
-    require('hjdivad/index').goto_tmux_session('todos', 'reference')
+    require('hjdivad/tmux').goto_tmux_session('todos', 'reference')
   end, { desc = 'tmux: reference', })
   vim.keymap.set('n', '<leader>tj', function()
-    require('hjdivad/index').goto_tmux_session('todos', 'journal')
+    require('hjdivad/tmux').goto_tmux_session('todos', 'journal')
   end, { desc = 'tmux: journal', })
 
   ---Mapping for opening up a generic REPL buffer + terminal pair.
   ---In general it's recommended to overwrite this mapping per-project with a
   ---specific repl command and repl file, for example:
   ---```lua
-  ---vim.api.nvim_set_keymap('n', '<leader>re', [[<cmd>ha.edit_repl('yarn repl', 'ts')]]<cr>')
+  ---vim.api.nvim_set_keymap('n', '<leader>re', function() require('hjdivad').edit_repl('yarn repl', 'ts') end)
   ---```
   vim.keymap.set('n', '<leader>re', function()
-    require('hjdivad/index').edit_generic_repl()
+    require('hjdivad').edit_generic_repl()
   end, { desc = 'Open a REPL in a terminal and a linked REPL input buffer', })
   vim.keymap.set('v', '<leader>re', ':TREPLSendSelection', { desc = 'Send selection to the REPL', })
 
   vim.keymap.set('n', '<leader>rr', '<cmd>TestFile<cr>', { desc = 'Run Test File (vim-test)', })
   vim.keymap.set('n', '<leader>rt', '<cmd>TestNearest<cr>', { desc = 'Run Nearest Test (vim-test)', })
   vim.keymap.set('n', '<leader>rd', function()
-    require('hjdivad/index').debug_nearest()
+    require('hjdivad/testing').debug_nearest()
   end, { desc = 'Debug Nearest Test (vim-test + hjdivad debug transform)', })
 
   --- yank file [path] to clipboard, using the best register we have available
@@ -356,8 +459,6 @@ local function setup_language_servers()
 end
 
 local function setup_plugins(config)
-  local hjdivad = require('hjdivad/index')
-
   require('malleatus').setup {}
 
   local function check(plugin)
@@ -370,7 +471,7 @@ local function setup_plugins(config)
     -- disable automapping from neoterm
     vim.g.neoterm_automap_keys = '😡😡STUPID_PLUGIN_DO_NOT_AUTOMAP'
 
-    hjdivad.setup_repls()
+    setup_repls()
   end
 
   if check('gitgutter') then
@@ -587,7 +688,7 @@ local function setup_plugins(config)
   end
 
   if check('vim-test') then
-    hjdivad.setup_vimtest()
+    require('hjdivad/testing').setup_vimtest()
   end
 end
 
@@ -608,7 +709,7 @@ local function create_user_commands()
   end, { desc = 'Update or install plugins' })
 
   vim.api.nvim_create_user_command('HiResize', function()
-    require('hjdivad/index').resize_with_terminal()
+    require('hjdivad/terminal').resize_with_terminal()
   end, { desc = 'Resize windows, ensuring a fixed-width terminal has appropriate width' })
 
   vim.api.nvim_create_user_command('HiDisableMarkdownFolding', function()
@@ -636,7 +737,6 @@ end
 local function main(config)
   --TODO: bootstrap here
 
-  local hjdivad = require('hjdivad/index')
   local opts = vim.tbl_deep_extend('force', {
     plugins = {},
     mappings = false,
@@ -650,8 +750,8 @@ local function main(config)
   setup_clipboard()
   setup_window_management()
   setup_statusline()
-  hjdivad.setup_terminal()
-  hjdivad.run_exrc()
+  require('hjdivad/terminal').setup_terminal()
+  require('hjdivad/exrc').run_exrc()
   create_user_commands()
 
   if opts.mappings then
@@ -660,6 +760,14 @@ local function main(config)
 end
 
 return {
+  -- lib entry points
+
+  edit_repl = edit_repl,
+  edit_generic_repl = edit_generic_repl,
+  toggle_nvim_tree = toggle_nvim_tree,
+
+  -- main entry points
+
   main = main,
   create_debug_functions = create_debug_functions,
 }
